aldous broder

+++ src/constants.rs
@@ -1,2 +1,2 @@
 pub const CELL_SIZE: f32 = 10.0;
-pub const GRID_SIZE: u8 = 20;
+pub const GRID_SIZE: usize = 20;
