+++ src/distances.rs
@@ -1,40 +1,39 @@
 use crate::grid::Grid;
 use std::collections::VecDeque;
 
-pub fn distances(grid: &mut Grid, start: (u8, u8)) {
-    let mut queue: VecDeque<(u8, u8)> = VecDeque::new();
+pub fn distances(grid: &mut Grid, start: (usize, usize)) {
+    let mut queue: VecDeque<(usize, usize)> = VecDeque::new();
 
     for i in 0..grid.size {
         for j in 0..grid.size {
-            grid.cells[i as usize][j as usize].distance = 0;
-            grid.cells[i as usize][j as usize].visited = false;
+            grid.cells[i][j].distance = 0;
+            grid.cells[i][j].visited = false;
         }
     }
 
     queue.push_back(start);
     while !queue.is_empty() {
         let (r, c) = queue.pop_front().unwrap();
-        grid.cells[r as usize][c as usize].visited = true;
-        grid.cells[r as usize][c as usize].distance += 1;
+        grid.cells[r][c].visited = true;
+        grid.cells[r][c].distance += 1;
         let neighbours = grid.neighbours((r, c));
         for n in neighbours {
-            if !grid.cells[n.0 as usize][n.1 as usize].visited {
+            if !grid.cells[n.0][n.1].visited {
                 queue.push_back((n.0, n.1));
-                grid.cells[n.0 as usize][n.1 as usize].distance =
-                    grid.cells[r as usize][c as usize].distance;
+                grid.cells[n.0][n.1].distance = grid.cells[r][c].distance;
             }
         }
     }
 
-    // let mut breadcrumbs: Vec<(u8, u8)> = vec![];
-    // grid.cells[target.0 as usize][target.1 as usize].solution_path = true;
+    // let mut breadcrumbs: Vec<(usize, usize)> = vec![];
+    // grid.cells[target.0][target.1].solution_path = true;
     // while target != start {
     //     let neighbours = grid.neighbours((target.0, target.1));
     //     for n in neighbours {
-    //         if grid.cells[n.0 as usize][n.1 as usize].distance
-    //             == grid.cells[target.0 as usize][target.1 as usize].distance - 1
+    //         if grid.cells[n.0][n.1].distance
+    //             == grid.cells[target.0][target.1].distance - 1
     //         {
-    //             grid.cells[n.0 as usize][n.1 as usize].solution_path = true;
+    //             grid.cells[n.0][n.1].solution_path = true;
     //             breadcrumbs.push((n.0, n.1));
     //             target = (n.0, n.1);
     //             break;
@@ -43,22 +42,22 @@ pub fn distances(grid: &mut Grid, start: (u8, u8)) {
     // }
 }
 
-pub fn solution(grid: &mut Grid, start: (u8, u8), end: (u8, u8)) {
-    let mut breadcrumbs: Vec<(u8, u8)> = vec![];
+pub fn solution(grid: &mut Grid, start: (usize, usize), end: (usize, usize)) {
+    let mut breadcrumbs: Vec<(usize, usize)> = vec![];
     let mut target = end;
     for i in 0..grid.size {
         for j in 0..grid.size {
-            grid.cells[i as usize][j as usize].solution_path = false;
+            grid.cells[i][j].solution_path = false;
         }
     }
-    grid.cells[target.0 as usize][target.1 as usize].solution_path = true;
+    grid.cells[target.0][target.1].solution_path = true;
     while target != start {
         let neighbours = grid.neighbours((target.0, target.1));
         for n in neighbours {
-            if grid.cells[n.0 as usize][n.1 as usize].distance
-                == grid.cells[target.0 as usize][target.1 as usize].distance - 1
+            if grid.cells[n.0][n.1].distance as i32
+                == grid.cells[target.0][target.1].distance as i32 - 1
             {
-                grid.cells[n.0 as usize][n.1 as usize].solution_path = true;
+                grid.cells[n.0][n.1].solution_path = true;
                 breadcrumbs.push((n.0, n.1));
                 target = (n.0, n.1);
                 break;
@@ -67,13 +66,13 @@ pub fn solution(grid: &mut Grid, start: (u8, u8), end: (u8, u8)) {
     }
 }
 
-pub fn farthest_cell(grid: &Grid, start: (u8, u8)) -> (u32, (u8, u8)) {
+pub fn farthest_cell(grid: &Grid, start: (usize, usize)) -> (u32, (usize, usize)) {
     let mut max = 1;
     let mut cell = start;
     for i in 0..grid.size {
         for j in 0..grid.size {
-            if grid.cells[i as usize][j as usize].distance > max {
-                max = grid.cells[i as usize][j as usize].distance;
+            if grid.cells[i][j].distance > max {
+                max = grid.cells[i][j].distance;
                 cell = (i, j);
             }
         }
